only need a reduction over floats
Ignore-this: 58e8921d7fdc286c086ed014dc6f1bd1

darcs-hash:20090911082208-9241b-7488d2e38ed0395a27dcc56593a778bc32440ddb.gz

+++ kernels/kernels.h
@@ -31,7 +31,7 @@ addIons
 // -----------------------------------------------------------------------------
 // Prelude
 // -----------------------------------------------------------------------------
-int  reducePlusf(float *xs, float N);
+float reducePlusf(float *xs, int N);
 
 void scanl1Plusi(int *in, int *out, int N);
 void scanr1Plusi(int *in, int *out, int N);
